requirements.txt
PyYAML
numpy==1.18
Pillow
opencv-python
scikit-learn
tensorflow==2.3